Show err msg when resend reminder to AD contact (#976)
* Show err msg when resend reminder to empty contact

https://eaflood.atlassian.net/browse/RUBY-1168

Currently if you click the 'Resend renewal reminder' link on a registration with an empty contact email it will just say 'Renewal sent to'.

With this message you could be mistaken for thinking it sent successfully. Or if you spot it, it wouldn;t be immediately obvious what the issue is.

We now have a design to display an error message if the link is clicked and the contact email is blank. This change implements it.

* Show err msg when resend reminder to AD contact

https://eaflood.atlassian.net/browse/RUBY-1169

Currently, if you click the 'Resend renewal reminder' link on a registration with the assisted digital contact email it will send the email and display the success message 'Renewal email sent to nccc-carrierbroker@environment-agency.gov.uk'.

NCCC know to use this email address when handling an AD registration where the user has no email address at all. To avoid getting emails sent to them mistakenly they have asked us to not send the email and show an error instead if the contact email matches the AD one.

* Refactor renewal_mailer spec

Refactoring the tests in preparation for adding new ones.

* Add new exception

Add a new exception to our exceptions module specifically for when the registrations contact email matches the assisted digital one.

* Implement check for assisted digital email match

* Update ctrl to handle AssistedDigitalContactEmailError

If the RenewalReminderMailer throws AssistedDigitalContactEmailError the controller will now be able to handle it and show a relevant error message.