Send reminder letters via Notify (#1071)
https://eaflood.atlassian.net/browse/RUBY-1288

This PR adds a service that sends reminder letters via GOV.UK Notify.

This is the first step in allowing us to remove the functionality to generate PDF letters overnight. We can test this with `bundle exec rake notify:test:ad_renewal_letter`

This also includes some testing changes:

Previously this relied on a lot of logic within the Order and OrderItem classes themselves, which was causing issues during testing when we wanted to check which methods were being called. This whole area still needs improvement though.

This also means a return for VCR so we can keep a copy of a Notify request for testing purposes.