Fix the Registration Transfer email validation (#1118)
https://eaflood.atlassian.net/browse/RUBY-1431

- When a user entered an invalid email/confirmation email pair, the web page
returned the expected error, but the input email was saved to the account_email
and the registration transfer emails would have been sent.

- This happened because the RegistrationTransferService was invoked as part of
the validation chain. When the email validation failed, there was nothing
to stop the RegistrationTransferService running.

- I refactored the RegistrationTransferForm to only call the service if the
model validations are successful. In the `submit` method there was a call to
`super` with empty attributes but it's not clear what this was meant to do.