Show correct user status in user export CSV (#1483)
The status precedence has been incorrect in #1482. With this commit we now
correctly shows the active/deactivated status as well as uses the presence of an
invitation token to determine if an active user's invitation is still pending,
being now in line with what's shown on the users index page.