Feature/ruby 2859 wcr ensure that scheduled job start and completion events are being recorded in logs (#1965)
* [RUBY-2859] adding a task that would run rake task and format the output

* [RUBY-2859] feat: Add timestamp to task start and finish messages

This commit adds a timestamp to the task start and finish messages in the `rake_and_format` script. The timestamp is formatted as `YYYY-MM-DD HH:MM:SS` using the `DateTime.now.strftime` method. This change improves the readability and provides more context for the task execution.

* [RUBY-2859] chore: Remove unnecessary blank line in schedule.rb file

* [RUBY-2859] feat: Update schedule.rb to use rake_and_format instead of rake

This commit modifies the schedule.rb file to use the rake_and_format task instead of the rake task. This change ensures that the correct tasks are executed when running the schedule.

* [RUBY-2859] feat: Remove unnecessary rake tasks

The commit removes the unnecessary rake tasks `one` and `two` from the `test.rake` file. These tasks were used for test purposes and no longer needed